requirements.txt
# Core dependencies
playwright>=1.40.0
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# UI dependencies
PyQt6>=6.4.0
ttkthemes>=3.2.0

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-playwright>=0.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.3.0

# Documentation
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0 